# docs/swagger.yaml
basePath: /api
definitions:
...
  dtos.CreateEmployeeDTO:
    properties:
      address:
        maxLength: 200
        minLength: 5
        type: string
      bi:
        type: string
      contract_type:
        enum:
        - permanent
        - temporary
        - intern
        - consultant
        type: string
      date_of_birth:
        type: string
      department_id:
        type: string
      district_id:
        type: string
      email:
        type: string
      full_name:
        maxLength: 100
        minLength: 3
        type: string
      gender:
        enum:
        - masculino
        - feminino
        type: string
      hiring_date:
        type: string
      iban:
        maxLength: 34
        minLength: 10
        type: string
      marital_status:
        enum:
        - single
        - married
        - divorced
        - widowed
        type: string
      nationality:
        maxLength: 50
        minLength: 2
        type: string
      phone_number:
        type: string
      position_id:
        type: string
      salary:
        type: number
      social_security:
        maxLength: 12
        minLength: 6
        type: string
      supervisor_id:
        type: string
    required:
    - address
    - bi
    - contract_type
    - date_of_birth
    - department_id
    - district_id
    - email
    - full_name
    - gender
    - hiring_date
    - iban
    - marital_status
    - nationality
    - phone_number
    - position_id
    - salary
    - social_security
    type: object
...
  dtos.EmployeeResponseDTO:
    properties:
      address:
        type: string
      bi:
        type: string
      contract_type:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      department_id:
        type: string
      district_id:
        type: string
      email:
        type: string
      employee_number:
        type: integer
      full_name:
        type: string
      gender:
        type: string
      hiring_date:
        type: string
      iban:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      marital_status:
        type: string
      nationality:
        type: string
      phone_number:
        type: string
      position_id:
        type: string
      salary:
        type: number
      social_security:
        type: string
      supervisor_id:
        type: string
      updated_at:
        type: string
    type: object
...
  dtos.UpdateEmployeeDTO:
    properties:
      address:
        maxLength: 200
        minLength: 5
        type: string
      contract_type:
        enum:
        - permanent
        - temporary
        - intern
        - consultant
        type: string
      department_id:
        type: string
      district_id:
        type: string
      email:
        type: string
      iban:
        maxLength: 34
        minLength: 10
        type: string
      is_active:
        type: boolean
      marital_status:
        enum:
        - single
        - married
        - divorced
        - widowed
        type: string
      phone_number:
        type: string
      position_id:
        type: string
      salary:
        type: number
      supervisor_id:
        type: string
    required:
    - district_id
    type: object
...
  utils.FieldError:
    properties:
      field:
        type: string
      label:
        type: string
      message:
        type: string
      tag:
        type: string
      value: {}
    type: object
  utils.Payload:
    properties:
      error:
        description: 'ex: validation_failed, conflict, internal'
        type: string
      fields:
        description: detalhes por campo
        items:
          $ref: '#/definitions/utils.FieldError'
        type: array
      message:
        description: resumo humanizado
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: suporte@empresa.com
    name: Equipa de Desenvolvimento
  description: API de Gestão de Recursos Humanos (Províncias, Municípios, Departamentos,
    etc.)
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://localhost:8080/terms
  title: HRMS API
  version: "1.0"
paths:
...
  /employees:
    get:
      description: Retorna uma lista paginada de funcionários
      parameters:
      - description: Número máximo de registros a retornar
        in: query
        name: limit
        type: integer
      - description: Número de registros a ignorar (para paginação)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.EmployeeResponseDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Lista funcionários
      tags:
      - Funcionários
    post:
      consumes:
      - application/json
      description: Registra um novo funcionário no sistema com todos os dados obrigatórios
      parameters:
      - description: Dados do funcionário
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateEmployeeDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.EmployeeResponseDTO'
        "400":
          description: Validação dos campos (Fields detalhados)
          schema:
            $ref: '#/definitions/utils.Payload'
        "409":
          description: Conflito (chave única)
          schema:
            $ref: '#/definitions/utils.Payload'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Cria um novo funcionário
      tags:
      - Funcionários
...
  /employees/{id}:
    delete:
      description: Exclui um funcionário pelo ID
      parameters:
      - description: ID do funcionário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Payload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Payload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Remove um funcionário
      tags:
      - Funcionários
    get:
      description: Retorna os dados de um funcionário específico
      parameters:
      - description: ID do funcionário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.EmployeeResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Payload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Payload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Busca um funcionário pelo ID
      tags:
      - Funcionários
    put:
      consumes:
      - application/json
      description: Modifica os dados de um funcionário existente
      parameters:
      - description: ID do funcionário
        in: path
        name: id
        required: true
        type: string
      - description: Dados atualizados
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateEmployeeDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.EmployeeResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Payload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Payload'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Payload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Atualiza os dados de um funcionário
      tags:
      - Funcionários
...
  /employees/search:
    get:
      consumes:
      - application/json
      description: Busca funcionários com texto, filtros e paginação
      parameters:
      - description: Texto de busca (nome, email, nº funcionário, etc.)
        in: query
        name: search
        type: string
      - description: 'Filtro avançado (ex.: dept:UUID, pos:UUID, ativo:true)'
        in: query
        name: filter
        type: string
      - description: Número máximo de registros a retornar
        in: query
        name: limit
        type: integer
      - description: Número de registros a ignorar (para paginação)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.EmployeeResponseDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Buscar funcionários
      tags:
      - Funcionários
...
schemes:
- http
swagger: "2.0"